вывод вопроса

+++ mobile_service/src/main/java/com/example/mobile/repos/QuestionRepos.java
@@ -3,5 +3,5 @@ package com.example.mobile.repos;
 import com.example.mobile.domain.Question;
 import org.springframework.data.repository.CrudRepository;
 
-public interface QuestionRepos extends CrudRepository<Question, Long> {
+public interface QuestionRepos extends CrudRepository<Question, Integer> {
 }
